chore: move register to library named file
