+++ bricks/fluttium_action/__brick__/{{name.snakeCase()}}/lib/{{name.snakeCase()}}.dart
@@ -1,0 +1,12 @@
+import 'package:fluttium/fluttium.dart';
+import 'package:{{name.snakeCase()}}/src/register.dart';
+
+export 'src/{{name.snakeCase()}}.dart';
+
+/// Will be executed by Fluttium on startup.
+void register(Registry registry) {
+    '{{name.camelCase()}}',
+    {{name.pascalCase()}}.new,
+    shortHandIs: #text,
+  );
+}
