+++ example/actions/expect_environment_text/lib/expect_environment_text.dart
@@ -1,0 +1,12 @@
+import 'package:expect_environment_text/expect_environment_text.dart';
+import 'package:fluttium/fluttium.dart';
+
+export 'src/expect_environment_text.dart';
+
+/// Will be executed by Fluttium on startup.
+void register(Registry registry) {
+  registry.registerAction(
+    'expectEnvironmentText',
+    ExpectEnvironmentText.new,
+  );
+}
