+++ packages/fluttium_cli/lib/src/bundles/fluttium_action_bundle.dart
@@ -6,9 +6,15 @@ import 'package:mason/mason.dart';
 final fluttiumActionBundle = MasonBundle.fromJson(<String, dynamic>{
   "files": [
     {
-      "path": "{{name.snakeCase()}}/lib/action.dart",
+      "path": "{{name.snakeCase()}}/lib/src/{{name.snakeCase()}}.dart",
       "data":
-          "aW1wb3J0ICdwYWNrYWdlOmZsdXR0aXVtL2ZsdXR0aXVtLmRhcnQnOwoKLy8vIHtAdGVtcGxhdGUge3tuYW1lLnNuYWtlQ2FzZSgpfX19Ci8vLyB7e3tkZXNjcmlwdGlvbn19fQovLy8KLy8vIEFuIGFjdGlvbiBpcyBpbW11dGFibGUgYW5kIGl0IHJlcHJlc2VudHMgYSBzdGVwIGluIHRoZSB1c2VyIGZsb3cuCi8vLwovLy8gVGhpcyBhY3Rpb24gY2FuIGJlIGludm9rZWQgZWl0aGVyIHVzaW5nIHRoZSBzaG9ydC1oYW5kIHZlcnNpb246Ci8vLwovLy8gYGBgeWFtbAovLy8gLSB7e25hbWUuY2FtZWxDYXNlKCl9fTogIkhlbGxvIFdvcmxkIgovLy8gYGBgCi8vLwovLy8gT3IgdXNpbmcgdGhlIHZlcmJvc2UgdmVyc2lvbjoKLy8vCi8vLyBgYGB5YW1sCi8vLyAtIHt7bmFtZS5jYW1lbENhc2UoKX19OgovLy8gICAgIHRleHQ6ICJIZWxsbyBXb3JsZCIKLy8vIGBgYAovLy8ge0BlbmR0ZW1wbGF0ZX0KY2xhc3Mge3tuYW1lLnBhc2NhbENhc2UoKX19IGV4dGVuZHMgQWN0aW9uIHsKICAvLy8ge0BtYWNybyB7e25hbWUuc25ha2VDYXNlKCl9fX0KICBjb25zdCB7e25hbWUucGFzY2FsQ2FzZSgpfX0oewogICAgcmVxdWlyZWQgdGhpcy50ZXh0LAogIH0pOwoKICBmaW5hbCBTdHJpbmc/IHRleHQ7CgogIC8vLyBDYWxsZWQgd2hlbiBpdCBleGVjdXRlcyB0aGUgYWN0aW9uIGluIGEgZmxvdyBmaWxlLgogIEBvdmVycmlkZQogIEZ1dHVyZTxib29sPiBleGVjdXRlKFRlc3RlciB0ZXN0ZXIpIGFzeW5jIHsKICAgIGlmICh0ZXh0ID09IG51bGwpIHsKICAgICAgcmV0dXJuIGZhbHNlOwogICAgfQoKICAgIGlmICghYXdhaXQgRXhwZWN0VmlzaWJsZSh0ZXh0OiB0ZXh0ISkuZXhlY3V0ZSh0ZXN0ZXIpKSB7CiAgICAgIHJldHVybiBmYWxzZTsKICAgIH0KICAgIHJldHVybiB0cnVlOwogIH0KCiAgQG92ZXJyaWRlCiAgU3RyaW5nIGRlc2NyaXB0aW9uKCkgPT4gJ3t7bmFtZS5zZW50ZW5jZUNhc2UoKX19ICIkdGV4dCInOwp9CgovLy8gV2lsbCBiZSBleGVjdXRlZCBieSBGbHV0dGl1bSBvbiBzdGFydHVwLgp2b2lkIHJlZ2lzdGVyKFJlZ2lzdHJ5IHJlZ2lzdHJ5KSB7CiAgcmVnaXN0cnkucmVnaXN0ZXJBY3Rpb24oCiAgICAne3tuYW1lLmNhbWVsQ2FzZSgpfX0nLAogICAge3tuYW1lLnBhc2NhbENhc2UoKX19Lm5ldywKICAgIHNob3J0SGFuZElzOiAjdGV4dCwKICApOwp9Cg==",
+          "aW1wb3J0ICdwYWNrYWdlOmZsdXR0aXVtL2ZsdXR0aXVtLmRhcnQnOwoKLy8vIHtAdGVtcGxhdGUge3tuYW1lLnNuYWtlQ2FzZSgpfX19Ci8vLyB7e3tkZXNjcmlwdGlvbn19fQovLy8KLy8vIEFuIGFjdGlvbiBpcyBpbW11dGFibGUgYW5kIGl0IHJlcHJlc2VudHMgYSBzdGVwIGluIHRoZSB1c2VyIGZsb3cuCi8vLwovLy8gVGhpcyBhY3Rpb24gY2FuIGJlIGludm9rZWQgZWl0aGVyIHVzaW5nIHRoZSBzaG9ydC1oYW5kIHZlcnNpb246Ci8vLwovLy8gYGBgeWFtbAovLy8gLSB7e25hbWUuY2FtZWxDYXNlKCl9fTogIkhlbGxvIFdvcmxkIgovLy8gYGBgCi8vLwovLy8gT3IgdXNpbmcgdGhlIHZlcmJvc2UgdmVyc2lvbjoKLy8vCi8vLyBgYGB5YW1sCi8vLyAtIHt7bmFtZS5jYW1lbENhc2UoKX19OgovLy8gICAgIHRleHQ6ICJIZWxsbyBXb3JsZCIKLy8vIGBgYAovLy8ge0BlbmR0ZW1wbGF0ZX0KY2xhc3Mge3tuYW1lLnBhc2NhbENhc2UoKX19IGV4dGVuZHMgQWN0aW9uIHsKICAvLy8ge0BtYWNybyB7e25hbWUuc25ha2VDYXNlKCl9fX0KICBjb25zdCB7e25hbWUucGFzY2FsQ2FzZSgpfX0oewogICAgcmVxdWlyZWQgdGhpcy50ZXh0LAogIH0pOwoKICBmaW5hbCBTdHJpbmc/IHRleHQ7CgogIC8vLyBDYWxsZWQgd2hlbiBpdCBleGVjdXRlcyB0aGUgYWN0aW9uIGluIGEgZmxvdyBmaWxlLgogIEBvdmVycmlkZQogIEZ1dHVyZTxib29sPiBleGVjdXRlKFRlc3RlciB0ZXN0ZXIpIGFzeW5jIHsKICAgIGlmICh0ZXh0ID09IG51bGwpIHsKICAgICAgcmV0dXJuIGZhbHNlOwogICAgfQoKICAgIGlmICghYXdhaXQgRXhwZWN0VmlzaWJsZSh0ZXh0OiB0ZXh0ISkuZXhlY3V0ZSh0ZXN0ZXIpKSB7CiAgICAgIHJldHVybiBmYWxzZTsKICAgIH0KICAgIHJldHVybiB0cnVlOwogIH0KCiAgQG92ZXJyaWRlCiAgU3RyaW5nIGRlc2NyaXB0aW9uKCkgPT4gJ3t7bmFtZS5zZW50ZW5jZUNhc2UoKX19ICIkdGV4dCInOwp9Cg==",
+      "type": "text"
+    },
+    {
+      "path": "{{name.snakeCase()}}/lib/{{name.snakeCase()}}.dart",
+      "data":
+          "aW1wb3J0ICdwYWNrYWdlOmZsdXR0aXVtL2ZsdXR0aXVtLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6e3tuYW1lLnNuYWtlQ2FzZSgpfX0vc3JjL3JlZ2lzdGVyLmRhcnQnOwoKZXhwb3J0ICdzcmMve3tuYW1lLnNuYWtlQ2FzZSgpfX0uZGFydCc7CgovLy8gV2lsbCBiZSBleGVjdXRlZCBieSBGbHV0dGl1bSBvbiBzdGFydHVwLgp2b2lkIHJlZ2lzdGVyKFJlZ2lzdHJ5IHJlZ2lzdHJ5KSB7CiAgICAne3tuYW1lLmNhbWVsQ2FzZSgpfX0nLAogICAge3tuYW1lLnBhc2NhbENhc2UoKX19Lm5ldywKICAgIHNob3J0SGFuZElzOiAjdGV4dCwKICApOwp9Cg==",
       "type": "text"
     },
     {
