+++ packages/fluttium_cli/test/fixtures/new/action/my_action/lib/my_action.dart
@@ -1,0 +1,12 @@
+import 'package:fluttium/fluttium.dart';
+import 'package:my_action/src/register.dart';
+
+export 'src/my_action.dart';
+
+/// Will be executed by Fluttium on startup.
+void register(Registry registry) {
+    'myAction',
+    MyAction.new,
+    shortHandIs: #text,
+  );
+}
