+++ packages/fluttium_driver/lib/src/bundles/fluttium_test_runner_bundle.dart
@@ -8,7 +8,7 @@ final fluttiumTestRunnerBundle = MasonBundle.fromJson(<String, dynamic>{
     {
       "path": "lib/fluttium_test_runner.dart",
       "data":
-          "aW1wb3J0ICdwYWNrYWdlOmZsdXR0ZXIvd2lkZ2V0cy5kYXJ0JyBoaWRlIEFjdGlvbjsKaW1wb3J0ICdwYWNrYWdlOmZsdXR0aXVtL2ZsdXR0aXVtLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6Zmx1dHRpdW1faW50ZXJmYWNlcy9mbHV0dGl1bV9pbnRlcmZhY2VzLmRhcnQnOwp7eyNhY3Rpb25zfX0KaW1wb3J0ICdwYWNrYWdlOnt7bmFtZX19L2FjdGlvbi5kYXJ0JyBhcyB7e25hbWUuc25ha2VDYXNlKCl9fTt7ey9hY3Rpb25zfX0KCkZ1dHVyZTx2b2lkPiBydW4oV2lkZ2V0c0JpbmRpbmcgYmluZGluZykgYXN5bmN7CiAgZmluYWwgcmVnaXN0cnkgPSBSZWdpc3RyeSgpO3t7I2FjdGlvbnN9fQogIHt7bmFtZS5zbmFrZUNhc2UoKX19LnJlZ2lzdGVyKHJlZ2lzdHJ5KTt7ey9hY3Rpb25zfX0KCiAgZmluYWwgdGVzdGVyID0gVGVzdGVyKGJpbmRpbmcsIHJlZ2lzdHJ5KTsKICBmaW5hbCBhY3Rpb25zID0gYXdhaXQgdGVzdGVyLmNvbnZlcnQoW3t7I3N0ZXBzfX0KICAgIFVzZXJGbG93U3RlcC5mcm9tSnNvbih7e3tzdGVwfX19KSx7ey9zdGVwc319CiAgXSk7CgogIGZvciAoZmluYWwgYWN0aW9uIGluIGFjdGlvbnMpIHsKICAgIGF3YWl0IGFjdGlvbigpOwogIH0KfQo=",
+          "aW1wb3J0ICdwYWNrYWdlOmZsdXR0ZXIvd2lkZ2V0cy5kYXJ0JyBoaWRlIEFjdGlvbjsKaW1wb3J0ICdwYWNrYWdlOmZsdXR0aXVtL2ZsdXR0aXVtLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6Zmx1dHRpdW1faW50ZXJmYWNlcy9mbHV0dGl1bV9pbnRlcmZhY2VzLmRhcnQnOwp7eyNhY3Rpb25zfX0KaW1wb3J0ICdwYWNrYWdlOnt7bmFtZS5zbmFrZUNhc2UoKX19L3t7bmFtZS5zbmFrZUNhc2UoKX19LmRhcnQnIGFzIHt7bmFtZS5zbmFrZUNhc2UoKX19O3t7L2FjdGlvbnN9fQoKRnV0dXJlPHZvaWQ+IHJ1bihXaWRnZXRzQmluZGluZyBiaW5kaW5nKSBhc3luY3sKICBmaW5hbCByZWdpc3RyeSA9IFJlZ2lzdHJ5KCk7e3sjYWN0aW9uc319CiAge3tuYW1lLnNuYWtlQ2FzZSgpfX0ucmVnaXN0ZXIocmVnaXN0cnkpO3t7L2FjdGlvbnN9fQoKICBmaW5hbCB0ZXN0ZXIgPSBUZXN0ZXIoYmluZGluZywgcmVnaXN0cnkpOwogIGZpbmFsIGFjdGlvbnMgPSBhd2FpdCB0ZXN0ZXIuY29udmVydChbe3sjc3RlcHN9fQogICAgVXNlckZsb3dTdGVwLmZyb21Kc29uKHt7e3N0ZXB9fX0pLHt7L3N0ZXBzfX0KICBdKTsKCiAgZm9yIChmaW5hbCBhY3Rpb24gaW4gYWN0aW9ucykgewogICAgYXdhaXQgYWN0aW9uKCk7CiAgfQp9Cg==",
       "type": "text"
     },
     {
@@ -22,7 +22,7 @@ final fluttiumTestRunnerBundle = MasonBundle.fromJson(<String, dynamic>{
     {
       "path": "post_gen.dart",
       "data":
-          "aW1wb3J0ICdkYXJ0OmlvJzsKCmltcG9ydCAncGFja2FnZTptYXNvbi9tYXNvbi5kYXJ0JzsKCnR5cGVkZWYgUHJvY2Vzc1J1bm5lciA9IEZ1dHVyZTxQcm9jZXNzUmVzdWx0PiBGdW5jdGlvbigKICBTdHJpbmcgZXhlY3V0YWJsZSwKICBMaXN0PFN0cmluZz4gYXJndW1lbnRzLCB7CiAgU3RyaW5nIHdvcmtpbmdEaXJlY3RvcnksCiAgYm9vbCBydW5JblNoZWxsLAp9KTsKCi8vIGNvdmVyYWdlOmlnbm9yZS1zdGFydApGdXR1cmU8dm9pZD4gcnVuKEhvb2tDb250ZXh0IGNvbnRleHQpID0+IHBvc3RHZW4oY29udGV4dCk7Ci8vIGNvdmVyYWdlOmlnbm9yZS1lbmQKCkZ1dHVyZTx2b2lkPiBwb3N0R2VuKAogIEhvb2tDb250ZXh0IGNvbnRleHQsIHsKICBEaXJlY3Rvcnk/IGRpcmVjdG9yeSwKICBQcm9jZXNzUnVubmVyIHJ1blByb2Nlc3MgPSBQcm9jZXNzLnJ1biwKfSkgYXN5bmMgewogIGZpbmFsIGluc3RhbGxpbmdBY3Rpb25zID0gY29udGV4dC5sb2dnZXIucHJvZ3Jlc3MoJ0luc3RhbGxpbmcgYWN0aW9ucycpOwogIGZpbmFsIHJlc3VsdCA9IGF3YWl0IHJ1blByb2Nlc3MoCiAgICAnZmx1dHRlcicsCiAgICBbJ3B1YicsICdnZXQnXSwKICAgIHJ1bkluU2hlbGw6IHRydWUsCiAgICB3b3JraW5nRGlyZWN0b3J5OiBEaXJlY3RvcnkuY3VycmVudC5wYXRoLAogICk7CgogIGlmIChyZXN1bHQuZXhpdENvZGUgIT0gMCkgewogICAgdGhyb3cgRXhjZXB0aW9uKCdGYWlsZWQgdG8gaW5zdGFsbCBhY3Rpb25zOiAke3Jlc3VsdC5zdGRlcnJ9Jyk7CiAgfQogIGluc3RhbGxpbmdBY3Rpb25zLmNvbXBsZXRlKCk7Cn0K",
+          "aW1wb3J0ICdkYXJ0OmlvJzsKCmltcG9ydCAncGFja2FnZTptYXNvbi9tYXNvbi5kYXJ0JzsKCnR5cGVkZWYgUHJvY2Vzc1J1bm5lciA9IEZ1dHVyZTxQcm9jZXNzUmVzdWx0PiBGdW5jdGlvbigKICBTdHJpbmcgZXhlY3V0YWJsZSwKICBMaXN0PFN0cmluZz4gYXJndW1lbnRzLCB7CiAgU3RyaW5nIHdvcmtpbmdEaXJlY3RvcnksCiAgYm9vbCBydW5JblNoZWxsLAp9KTsKCi8vIGNvdmVyYWdlOmlnbm9yZS1zdGFydApGdXR1cmU8dm9pZD4gcnVuKEhvb2tDb250ZXh0IGNvbnRleHQpID0+IHBvc3RHZW4oY29udGV4dCk7Ci8vIGNvdmVyYWdlOmlnbm9yZS1lbmQKCkZ1dHVyZTx2b2lkPiBwb3N0R2VuKAogIEhvb2tDb250ZXh0IGNvbnRleHQsIHsKICBEaXJlY3Rvcnk/IGRpcmVjdG9yeSwKICBQcm9jZXNzUnVubmVyIHJ1blByb2Nlc3MgPSBQcm9jZXNzLnJ1biwKfSkgYXN5bmMgewogIGZpbmFsIGluc3RhbGxpbmdBY3Rpb25zID0gY29udGV4dC5sb2dnZXIucHJvZ3Jlc3MoJ0luc3RhbGxpbmcgYWN0aW9ucycpOwogIGZpbmFsIHJlc3VsdCA9IGF3YWl0IHJ1blByb2Nlc3MoCiAgICAnZmx1dHRlcicsCiAgICBbJ3B1YicsICdnZXQnXSwKICAgIHJ1bkluU2hlbGw6IHRydWUsCiAgICB3b3JraW5nRGlyZWN0b3J5OiBEaXJlY3RvcnkuY3VycmVudC5wYXRoLAogICk7CgogIGlmIChyZXN1bHQuZXhpdENvZGUgIT0gMCkgewogICAgaW5zdGFsbGluZ0FjdGlvbnMuZmFpbCgpOwogICAgdGhyb3cgRXhjZXB0aW9uKCdGYWlsZWQgdG8gaW5zdGFsbCBhY3Rpb25zOiAke3Jlc3VsdC5zdGRlcnJ9Jyk7CiAgfQogIGluc3RhbGxpbmdBY3Rpb25zLmNvbXBsZXRlKCk7Cn0K",
       "type": "text"
     },
     {
@@ -34,7 +34,7 @@ final fluttiumTestRunnerBundle = MasonBundle.fromJson(<String, dynamic>{
     {
       "path": "test/post_gen_test.dart",
       "data":
-          "aW1wb3J0ICdkYXJ0OmlvJzsKCmltcG9ydCAncGFja2FnZTptYXNvbi9tYXNvbi5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOm1vY2t0YWlsL21vY2t0YWlsLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6dGVzdC90ZXN0LmRhcnQnOwoKaW1wb3J0ICcuLi9wb3N0X2dlbi5kYXJ0JzsKCmNsYXNzIF9GYWtlSG9va0NvbnRleHQgZXh0ZW5kcyBGYWtlIGltcGxlbWVudHMgSG9va0NvbnRleHQgewogIF9GYWtlSG9va0NvbnRleHQoe0xvZ2dlcj8gbG9nZ2VyfSkgOiBfbG9nZ2VyID0gbG9nZ2VyID8/IF9Nb2NrTG9nZ2VyKCk7CgogIGZpbmFsIExvZ2dlciBfbG9nZ2VyOwoKICB2YXIgX3ZhcnMgPSA8U3RyaW5nLCBkeW5hbWljPnt9OwoKICBAb3ZlcnJpZGUKICBNYXA8U3RyaW5nLCBkeW5hbWljPiBnZXQgdmFycyA9PiBfdmFyczsKCiAgQG92ZXJyaWRlCiAgc2V0IHZhcnMoTWFwPFN0cmluZywgZHluYW1pYz4gdmFsdWUpID0+IF92YXJzID0gdmFsdWU7CgogIEBvdmVycmlkZQogIExvZ2dlciBnZXQgbG9nZ2VyID0+IF9sb2dnZXI7Cn0KCmNsYXNzIF9Nb2NrTG9nZ2VyIGV4dGVuZHMgTW9jayBpbXBsZW1lbnRzIExvZ2dlciB7fQoKY2xhc3MgX01vY2tQcm9ncmVzcyBleHRlbmRzIE1vY2sgaW1wbGVtZW50cyBQcm9ncmVzcyB7fQoKY2xhc3MgX01vY2tQcm9jZXNzUmVzdWx0IGV4dGVuZHMgTW9jayBpbXBsZW1lbnRzIFByb2Nlc3NSZXN1bHQge30KCnZvaWQgbWFpbigpIHsKICBncm91cCgncG9zdEdlbicsICgpIHsKICAgIGxhdGUgSG9va0NvbnRleHQgY29udGV4dDsKICAgIGxhdGUgTG9nZ2VyIGxvZ2dlcjsKICAgIGxhdGUgUHJvZ3Jlc3MgaW5zdGFsbGluZ0FjdGlvbnM7CgogICAgc2V0VXAoKCkgewogICAgICBsb2dnZXIgPSBfTW9ja0xvZ2dlcigpOwogICAgICBjb250ZXh0ID0gX0Zha2VIb29rQ29udGV4dChsb2dnZXI6IGxvZ2dlcik7CiAgICAgIGluc3RhbGxpbmdBY3Rpb25zID0gX01vY2tQcm9ncmVzcygpOwoKICAgICAgd2hlbigoKSA9PiBsb2dnZXIucHJvZ3Jlc3MoYW55KCkpKS50aGVuUmV0dXJuKGluc3RhbGxpbmdBY3Rpb25zKTsKICAgIH0pOwoKICAgIHRlc3QoJ2luc3RhbGxzIGFjdGlvbnMgY29ycmVjdGx5JywgKCkgYXN5bmMgewogICAgICB2YXIgcHJvY2Vzc1J1bm5lckNhbGxDb3VudCA9IDA7CgogICAgICBmaW5hbCByZXN1bHQgPSBfTW9ja1Byb2Nlc3NSZXN1bHQoKTsKICAgICAgd2hlbigoKSA9PiByZXN1bHQuZXhpdENvZGUpLnRoZW5SZXR1cm4oRXhpdENvZGUuc3VjY2Vzcy5jb2RlKTsKCiAgICAgIGF3YWl0IHBvc3RHZW4oCiAgICAgICAgY29udGV4dCwKICAgICAgICBydW5Qcm9jZXNzOiAoCiAgICAgICAgICBleGVjdXRhYmxlLAogICAgICAgICAgYXJncywgewogICAgICAgICAgU3RyaW5nPyB3b3JraW5nRGlyZWN0b3J5LAogICAgICAgICAgYm9vbD8gcnVuSW5TaGVsbCwKICAgICAgICB9KSBhc3luYyB7CiAgICAgICAgICBwcm9jZXNzUnVubmVyQ2FsbENvdW50Kys7CiAgICAgICAgICBleHBlY3QoZXhlY3V0YWJsZSwgZXF1YWxzKCdmbHV0dGVyJykpOwogICAgICAgICAgZXhwZWN0KGFyZ3MsIGVxdWFscyhbJ3B1YicsICdnZXQnXSkpOwogICAgICAgICAgZXhwZWN0KHdvcmtpbmdEaXJlY3RvcnksIGVxdWFscyhEaXJlY3RvcnkuY3VycmVudC5wYXRoKSk7CiAgICAgICAgICBleHBlY3QocnVuSW5TaGVsbCwgaXNUcnVlKTsKICAgICAgICAgIHJldHVybiByZXN1bHQ7CiAgICAgICAgfSwKICAgICAgKTsKCiAgICAgIGV4cGVjdChwcm9jZXNzUnVubmVyQ2FsbENvdW50LCBlcXVhbHMoMSkpOwogICAgICB2ZXJpZnkoCiAgICAgICAgKCkgPT4gbG9nZ2VyLnByb2dyZXNzKGFueSh0aGF0OiBlcXVhbHMoJ0luc3RhbGxpbmcgYWN0aW9ucycpKSksCiAgICAgICkuY2FsbGVkKDEpOwogICAgICB2ZXJpZnkoKCkgPT4gaW5zdGFsbGluZ0FjdGlvbnMuY29tcGxldGUoKSkuY2FsbGVkKDEpOwogICAgfSk7CgogICAgdGVzdCgnZmFpbHMgaW5zdGFsbGluZyBhY3Rpb25zIHRocm93cyBleGNlcHRpb24nLCAoKSBhc3luYyB7CiAgICAgIHZhciBwcm9jZXNzUnVubmVyQ2FsbENvdW50ID0gMDsKCiAgICAgIGZpbmFsIHJlc3VsdCA9IF9Nb2NrUHJvY2Vzc1Jlc3VsdCgpOwogICAgICB3aGVuKCgpID0+IHJlc3VsdC5leGl0Q29kZSkudGhlblJldHVybihFeGl0Q29kZS50ZW1wRmFpbC5jb2RlKTsKICAgICAgd2hlbigoKSA9PiByZXN1bHQuc3RkZXJyKS50aGVuUmV0dXJuKCdmYWtlIGVycm9yJyk7CgogICAgICB0cnkgewogICAgICAgIGF3YWl0IHBvc3RHZW4oCiAgICAgICAgICBjb250ZXh0LAogICAgICAgICAgcnVuUHJvY2VzczogKAogICAgICAgICAgICBleGVjdXRhYmxlLAogICAgICAgICAgICBhcmdzLCB7CiAgICAgICAgICAgIFN0cmluZz8gd29ya2luZ0RpcmVjdG9yeSwKICAgICAgICAgICAgYm9vbD8gcnVuSW5TaGVsbCwKICAgICAgICAgIH0pIGFzeW5jIHsKICAgICAgICAgICAgcHJvY2Vzc1J1bm5lckNhbGxDb3VudCsrOwogICAgICAgICAgICBleHBlY3QoZXhlY3V0YWJsZSwgZXF1YWxzKCdmbHV0dGVyJykpOwogICAgICAgICAgICBleHBlY3QoYXJncywgZXF1YWxzKFsncHViJywgJ2dldCddKSk7CiAgICAgICAgICAgIGV4cGVjdCh3b3JraW5nRGlyZWN0b3J5LCBlcXVhbHMoRGlyZWN0b3J5LmN1cnJlbnQucGF0aCkpOwogICAgICAgICAgICBleHBlY3QocnVuSW5TaGVsbCwgaXNUcnVlKTsKICAgICAgICAgICAgcmV0dXJuIHJlc3VsdDsKICAgICAgICAgIH0sCiAgICAgICAgKTsKICAgICAgfSBvbiBFeGNlcHRpb24gY2F0Y2ggKGVycikgewogICAgICAgIGV4cGVjdCgKICAgICAgICAgIGVyci50b1N0cmluZygpLAogICAgICAgICAgZXF1YWxzKCdFeGNlcHRpb246IEZhaWxlZCB0byBpbnN0YWxsIGFjdGlvbnM6IGZha2UgZXJyb3InKSwKICAgICAgICApOwogICAgICB9CgogICAgICBleHBlY3QocHJvY2Vzc1J1bm5lckNhbGxDb3VudCwgZXF1YWxzKDEpKTsKICAgICAgdmVyaWZ5KAogICAgICAgICgpID0+IGxvZ2dlci5wcm9ncmVzcyhhbnkodGhhdDogZXF1YWxzKCdJbnN0YWxsaW5nIGFjdGlvbnMnKSkpLAogICAgICApLmNhbGxlZCgxKTsKICAgICAgdmVyaWZ5TmV2ZXIoKCkgPT4gaW5zdGFsbGluZ0FjdGlvbnMuY29tcGxldGUoKSk7CiAgICB9KTsKICB9KTsKfQo=",
+          "aW1wb3J0ICdkYXJ0OmlvJzsKCmltcG9ydCAncGFja2FnZTptYXNvbi9tYXNvbi5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOm1vY2t0YWlsL21vY2t0YWlsLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6dGVzdC90ZXN0LmRhcnQnOwoKaW1wb3J0ICcuLi9wb3N0X2dlbi5kYXJ0JzsKCmNsYXNzIF9GYWtlSG9va0NvbnRleHQgZXh0ZW5kcyBGYWtlIGltcGxlbWVudHMgSG9va0NvbnRleHQgewogIF9GYWtlSG9va0NvbnRleHQoe0xvZ2dlcj8gbG9nZ2VyfSkgOiBfbG9nZ2VyID0gbG9nZ2VyID8/IF9Nb2NrTG9nZ2VyKCk7CgogIGZpbmFsIExvZ2dlciBfbG9nZ2VyOwoKICB2YXIgX3ZhcnMgPSA8U3RyaW5nLCBkeW5hbWljPnt9OwoKICBAb3ZlcnJpZGUKICBNYXA8U3RyaW5nLCBkeW5hbWljPiBnZXQgdmFycyA9PiBfdmFyczsKCiAgQG92ZXJyaWRlCiAgc2V0IHZhcnMoTWFwPFN0cmluZywgZHluYW1pYz4gdmFsdWUpID0+IF92YXJzID0gdmFsdWU7CgogIEBvdmVycmlkZQogIExvZ2dlciBnZXQgbG9nZ2VyID0+IF9sb2dnZXI7Cn0KCmNsYXNzIF9Nb2NrTG9nZ2VyIGV4dGVuZHMgTW9jayBpbXBsZW1lbnRzIExvZ2dlciB7fQoKY2xhc3MgX01vY2tQcm9ncmVzcyBleHRlbmRzIE1vY2sgaW1wbGVtZW50cyBQcm9ncmVzcyB7fQoKY2xhc3MgX01vY2tQcm9jZXNzUmVzdWx0IGV4dGVuZHMgTW9jayBpbXBsZW1lbnRzIFByb2Nlc3NSZXN1bHQge30KCnZvaWQgbWFpbigpIHsKICBncm91cCgncG9zdEdlbicsICgpIHsKICAgIGxhdGUgSG9va0NvbnRleHQgY29udGV4dDsKICAgIGxhdGUgTG9nZ2VyIGxvZ2dlcjsKICAgIGxhdGUgUHJvZ3Jlc3MgaW5zdGFsbGluZ0FjdGlvbnM7CgogICAgc2V0VXAoKCkgewogICAgICBsb2dnZXIgPSBfTW9ja0xvZ2dlcigpOwogICAgICBjb250ZXh0ID0gX0Zha2VIb29rQ29udGV4dChsb2dnZXI6IGxvZ2dlcik7CiAgICAgIGluc3RhbGxpbmdBY3Rpb25zID0gX01vY2tQcm9ncmVzcygpOwoKICAgICAgd2hlbigoKSA9PiBsb2dnZXIucHJvZ3Jlc3MoYW55KCkpKS50aGVuUmV0dXJuKGluc3RhbGxpbmdBY3Rpb25zKTsKICAgIH0pOwoKICAgIHRlc3QoJ2luc3RhbGxzIGFjdGlvbnMgY29ycmVjdGx5JywgKCkgYXN5bmMgewogICAgICB2YXIgcHJvY2Vzc1J1bm5lckNhbGxDb3VudCA9IDA7CgogICAgICBmaW5hbCByZXN1bHQgPSBfTW9ja1Byb2Nlc3NSZXN1bHQoKTsKICAgICAgd2hlbigoKSA9PiByZXN1bHQuZXhpdENvZGUpLnRoZW5SZXR1cm4oRXhpdENvZGUuc3VjY2Vzcy5jb2RlKTsKCiAgICAgIGF3YWl0IHBvc3RHZW4oCiAgICAgICAgY29udGV4dCwKICAgICAgICBydW5Qcm9jZXNzOiAoCiAgICAgICAgICBleGVjdXRhYmxlLAogICAgICAgICAgYXJncywgewogICAgICAgICAgU3RyaW5nPyB3b3JraW5nRGlyZWN0b3J5LAogICAgICAgICAgYm9vbD8gcnVuSW5TaGVsbCwKICAgICAgICB9KSBhc3luYyB7CiAgICAgICAgICBwcm9jZXNzUnVubmVyQ2FsbENvdW50Kys7CiAgICAgICAgICBleHBlY3QoZXhlY3V0YWJsZSwgZXF1YWxzKCdmbHV0dGVyJykpOwogICAgICAgICAgZXhwZWN0KGFyZ3MsIGVxdWFscyhbJ3B1YicsICdnZXQnXSkpOwogICAgICAgICAgZXhwZWN0KHdvcmtpbmdEaXJlY3RvcnksIGVxdWFscyhEaXJlY3RvcnkuY3VycmVudC5wYXRoKSk7CiAgICAgICAgICBleHBlY3QocnVuSW5TaGVsbCwgaXNUcnVlKTsKICAgICAgICAgIHJldHVybiByZXN1bHQ7CiAgICAgICAgfSwKICAgICAgKTsKCiAgICAgIGV4cGVjdChwcm9jZXNzUnVubmVyQ2FsbENvdW50LCBlcXVhbHMoMSkpOwogICAgICB2ZXJpZnkoCiAgICAgICAgKCkgPT4gbG9nZ2VyLnByb2dyZXNzKGFueSh0aGF0OiBlcXVhbHMoJ0luc3RhbGxpbmcgYWN0aW9ucycpKSksCiAgICAgICkuY2FsbGVkKDEpOwogICAgICB2ZXJpZnkoKCkgPT4gaW5zdGFsbGluZ0FjdGlvbnMuY29tcGxldGUoKSkuY2FsbGVkKDEpOwogICAgfSk7CgogICAgdGVzdCgnZmFpbHMgaW5zdGFsbGluZyBhY3Rpb25zIHRocm93cyBleGNlcHRpb24nLCAoKSBhc3luYyB7CiAgICAgIHZhciBwcm9jZXNzUnVubmVyQ2FsbENvdW50ID0gMDsKCiAgICAgIGZpbmFsIHJlc3VsdCA9IF9Nb2NrUHJvY2Vzc1Jlc3VsdCgpOwogICAgICB3aGVuKCgpID0+IHJlc3VsdC5leGl0Q29kZSkudGhlblJldHVybihFeGl0Q29kZS50ZW1wRmFpbC5jb2RlKTsKICAgICAgd2hlbigoKSA9PiByZXN1bHQuc3RkZXJyKS50aGVuUmV0dXJuKCdmYWtlIGVycm9yJyk7CgogICAgICB0cnkgewogICAgICAgIGF3YWl0IHBvc3RHZW4oCiAgICAgICAgICBjb250ZXh0LAogICAgICAgICAgcnVuUHJvY2VzczogKAogICAgICAgICAgICBleGVjdXRhYmxlLAogICAgICAgICAgICBhcmdzLCB7CiAgICAgICAgICAgIFN0cmluZz8gd29ya2luZ0RpcmVjdG9yeSwKICAgICAgICAgICAgYm9vbD8gcnVuSW5TaGVsbCwKICAgICAgICAgIH0pIGFzeW5jIHsKICAgICAgICAgICAgcHJvY2Vzc1J1bm5lckNhbGxDb3VudCsrOwogICAgICAgICAgICBleHBlY3QoZXhlY3V0YWJsZSwgZXF1YWxzKCdmbHV0dGVyJykpOwogICAgICAgICAgICBleHBlY3QoYXJncywgZXF1YWxzKFsncHViJywgJ2dldCddKSk7CiAgICAgICAgICAgIGV4cGVjdCh3b3JraW5nRGlyZWN0b3J5LCBlcXVhbHMoRGlyZWN0b3J5LmN1cnJlbnQucGF0aCkpOwogICAgICAgICAgICBleHBlY3QocnVuSW5TaGVsbCwgaXNUcnVlKTsKICAgICAgICAgICAgcmV0dXJuIHJlc3VsdDsKICAgICAgICAgIH0sCiAgICAgICAgKTsKICAgICAgfSBvbiBFeGNlcHRpb24gY2F0Y2ggKGVycikgewogICAgICAgIGV4cGVjdCgKICAgICAgICAgIGVyci50b1N0cmluZygpLAogICAgICAgICAgZXF1YWxzKCdFeGNlcHRpb246IEZhaWxlZCB0byBpbnN0YWxsIGFjdGlvbnM6IGZha2UgZXJyb3InKSwKICAgICAgICApOwogICAgICB9CgogICAgICBleHBlY3QocHJvY2Vzc1J1bm5lckNhbGxDb3VudCwgZXF1YWxzKDEpKTsKICAgICAgdmVyaWZ5KAogICAgICAgICgpID0+IGxvZ2dlci5wcm9ncmVzcyhhbnkodGhhdDogZXF1YWxzKCdJbnN0YWxsaW5nIGFjdGlvbnMnKSkpLAogICAgICApLmNhbGxlZCgxKTsKICAgICAgdmVyaWZ5KCgpID0+IGluc3RhbGxpbmdBY3Rpb25zLmZhaWwoKSkuY2FsbGVkKDEpOwogICAgICB2ZXJpZnlOZXZlcigoKSA9PiBpbnN0YWxsaW5nQWN0aW9ucy5jb21wbGV0ZSgpKTsKICAgIH0pOwogIH0pOwp9Cg==",
       "type": "text"
     }
   ],
